Remove libkolab and libcalendaring from plugins list
Summary:
As a follow-up to #4033, these two plugins itself do nothing, so
it's not needed to list them in enabled plugins list. They will
be automatically loaded when required by other plugins.

To be precise, this prevents libcalendaring js/css resources to be
loaded on logon page, where they are not needed.

Reviewers: #pykolab_developers, vanmeeuwen

Reviewed By: #pykolab_developers, vanmeeuwen

Subscribers: vanmeeuwen

Projects: #pykolab

Differential Revision: https://git.kolab.org/D100