Set default value of --max-tasks to None, which effectively makes pool workers work forever.
Summary: That should be a workaround for issue when multiprocessing spawns a process which never gets a task and that ends up with all pool processes not getting into a state to accept new task and emails start congest. Most probably that is a deadlock situation described in https://codewithoutrules.com/2018/09/04/python-multiprocessing/ . The solution is in Python3.

Reviewers: vanmeeuwen

Reviewed By: vanmeeuwen

Subscribers: petersen, machniak

Differential Revision: https://git.kolab.org/D803